pyproject.toml
[project]
name = "python-finance"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.6",
    "notebook>=7.4.5",
    "pandas>=2.3.2",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.2",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
